requirements.txt
tesserocr
numpy
Pillow
mss